Bump sbt 1.7.2 => 1.8.2 / add sbt-tpolecat and update the code accordingly

+++ modules/extras-scala-io/shared/src/test/scala/extras/scala/io/file/FileUtils.scala
@@ -11,24 +11,31 @@ object FileUtils {
 
   def cleanUpFilesInside(file: File): Unit = {
     @tailrec
-    def cleanAll(files: Seq[File]): Unit = files match {
-      case file +: rest =>
-        if (file.isDirectory) {
-          val list = file.listFiles
-          if (list.isEmpty) {
-            Try(file.delete()).foreach(_ => ())
+    def cleanAll(files: Seq[File]): Unit = {
+      if (files.isEmpty) ()
+      else {
+        val head = files.headOption
+        val rest = files.drop(1)
+        head match {
+          case Some(file) =>
+            if (file.isDirectory) {
+              val list = file.listFiles
+              if (list.isEmpty) {
+                Try(file.delete()).foreach(_ => ())
+                cleanAll(rest)
+              } else {
+                cleanAll((list.toVector ++ rest) :+ file)
+              }
+            } else {
+              if (file.exists()) {
+                Try(file.delete()).foreach(_ => ())
+                cleanAll(rest)
+              } else cleanAll(rest)
+            }
+          case None =>
             cleanAll(rest)
-          } else {
-            cleanAll((list.toVector ++ rest) :+ file)
-          }
-        } else {
-          if (file.exists()) {
-            Try(file.delete()).foreach(_ => ())
-            cleanAll(rest)
-          } else cleanAll(rest)
         }
-      case Seq() =>
-        ()
+      }
     }
     if (file.exists) {
       if (file.isDirectory) {
